requirements.txt
flask
requests
fastapi
uvicorn
pydantic
langgraph
langchain-deepseek
langchain-core
langchain-mcp-adapters
python-dotenv
httpx
mcp
apscheduler
aiosqlite
langgraph-checkpoint-sqlite
ddgs
